vxlan: decap use vlib_buffer_enqueue_to_next

+++ src/vnet/vxlan/vxlan.h
@@ -78,7 +78,7 @@ typedef struct
   u32 mcast_sw_if_index;
 
   /* decap next index */
-  u32 decap_next_index;
+  u16 decap_next_index;
 
   /* The FIB index for src/dst addresses */
   u32 encap_fib_index;
